requirements.txt
flask
langchain
langchain-community
langchain-google-genai
tavily-python
python-dotenv
